fix review comments and add testcases
