+++ test/data/files/test_ee.py
@@ -1,0 +1,3 @@
+import os
+
+print('this is execution environment {}'.format(os.environ['HOSTNAME']))
