+++ test/integration/test_interface.py
@@ -2,7 +2,8 @@ import os
 import pytest
 import shutil
 
-from ansible_runner.interface import run, run_async
+from ansible_runner.interface import run, run_async, run_command, run_command_async, get_plugin_docs, \
+    get_plugin_docs_async, get_plugin_list, get_ansible_config, get_inventory
 
 
 def test_run():
@@ -134,3 +135,183 @@ def test_inventory_absolute_path(test_data_dir):
     # hosts can be down-selected to one inventory out of those available
     assert 'host_1' in stdout
     assert 'host_2' not in stdout
+
+
+def test_run_command(test_data_dir):
+    private_data_dir = os.path.join(test_data_dir, 'debug')
+    inventory = os.path.join(private_data_dir, 'inventory', 'inv_1')
+    playbook = os.path.join(private_data_dir, 'project', 'debug.yml')
+    out, err = run_command(
+        private_data_dir=private_data_dir,
+        executable_cmd='ansible-playbook',
+        cmdline_args=[playbook, '-i', inventory]
+    )
+    assert "Hello world!" in out
+    assert err == ''
+
+
+def test_run_ansible_command_within_container(test_data_dir, container_runtime_installed):
+    private_data_dir = os.path.join(test_data_dir, 'debug')
+    inventory = os.path.join(private_data_dir, 'inventory', 'inv_1')
+    playbook = os.path.join(private_data_dir, 'project', 'debug.yml')
+    container_kwargs = {
+        'process_isolation_executable': container_runtime_installed,
+        'process_isolation': True,
+        'container_image': 'quay.io/ansible/ansible-runner:devel'
+    }
+    out, err = run_command(
+        private_data_dir=private_data_dir,
+        executable_cmd='ansible-playbook',
+        cmdline_args=[playbook, '-i', inventory],
+        **container_kwargs
+    )
+    assert "Hello world!" in out
+    assert err == ''
+
+
+def test_run_script_within_container(test_data_dir, container_runtime_installed):
+    private_data_dir = os.path.join(test_data_dir, 'debug')
+    script_path = os.path.join(test_data_dir, 'files')
+    container_volume_mounts = ["{}:{}:Z".format(script_path, script_path)]
+    container_kwargs = {
+        'process_isolation_executable': container_runtime_installed,
+        'process_isolation': True,
+        'container_image': 'quay.io/ansible/ansible-runner:devel',
+        'container_volume_mounts': container_volume_mounts
+    }
+    out, err = run_command(
+        private_data_dir=private_data_dir,
+        executable_cmd='python3',
+        cmdline_args=[os.path.join(script_path, 'test_ee.py')],
+        **container_kwargs
+    )
+
+    assert "this is execution environment" in out
+    assert err == ''
+
+
+def test_run_command_async(test_data_dir):
+    private_data_dir = os.path.join(test_data_dir, 'debug')
+    inventory = os.path.join(private_data_dir, 'inventory', 'inv_1')
+    playbook = os.path.join(private_data_dir, 'project', 'debug.yml')
+    thread, r = run_command_async(
+        private_data_dir=private_data_dir,
+        executable_cmd='ansible-playbook',
+        cmdline_args=[playbook, '-i', inventory]
+    )
+    thread.join()
+    out = r.stdout.read()
+    assert "Hello world!" in out
+    assert r.status == 'successful'
+
+
+def test_get_plugin_docs():
+    out, _ = get_plugin_docs(
+        plugin_names=['ansible.builtin.file', 'ansible.builtin.copy'],
+        plugin_type='module',
+        response_format='json',
+        quiet=True
+    )
+    assert out['ansible.builtin.copy'] is not None
+    assert out['ansible.builtin.file'] is not None
+
+
+def test_get_plugin_docs_async():
+    thread, r = get_plugin_docs_async(
+        plugin_names=['ansible.builtin.file', 'ansible.builtin.copy'],
+        plugin_type='module',
+        response_format='json',
+        quiet=True
+    )
+    thread.join()
+    out = r.stdout.read()
+    assert 'ansible.builtin.copy' in out
+    assert 'ansible.builtin.file' in out
+    assert r.status == 'successful'
+
+
+def test_get_plugin_docs_within_container(container_runtime_installed):
+    container_kwargs = {
+        'process_isolation_executable': container_runtime_installed,
+        'process_isolation': True,
+        'container_image': 'quay.io/ansible/ansible-runner:devel'
+    }
+    out, _ = get_plugin_docs(
+        plugin_names=['ansible.builtin.file', 'ansible.builtin.copy'],
+        plugin_type='module',
+        response_format='json',
+        quiet=True,
+        **container_kwargs
+    )
+    assert out['ansible.builtin.copy'] is not None
+    assert out['ansible.builtin.file'] is not None
+
+
+def test_get_plugin_docs_list():
+    out, _ = get_plugin_list(
+        list_files=True,
+        response_format='json',
+        quiet=True
+    )
+    assert out['copy'] is not None
+    assert out['file'] is not None
+
+
+def test_get_plugin_docs_list_within_container(container_runtime_installed):
+    container_kwargs = {
+        'process_isolation_executable': container_runtime_installed,
+        'process_isolation': True,
+        'container_image': 'quay.io/ansible/ansible-runner:devel'
+    }
+    out, _ = get_plugin_list(
+        list_files=True,
+        response_format='json',
+        quiet=True,
+        **container_kwargs
+    )
+    assert out['copy'] is not None
+    assert out['file'] is not None
+
+
+def test_ansible_config():
+    out, _ = get_ansible_config(
+        action='list',
+        quiet=True
+    )
+    assert 'DEFAULT_VERBOSITY' in out
+
+
+def test_get_inventory(test_data_dir):
+    private_data_dir = os.path.join(test_data_dir, 'debug')
+    inventory1 = os.path.join(private_data_dir, 'inventory', 'inv_1')
+    inventory2 = os.path.join(private_data_dir, 'inventory', 'inv_2')
+
+    out, _ = get_inventory(
+        action='list',
+        inventories=[inventory1, inventory2],
+        response_format='json',
+        quiet=True
+    )
+    assert 'host_1' in out['ungrouped']['hosts']
+    assert 'host_2' in out['ungrouped']['hosts']
+
+
+def test_get_inventory_within_container(test_data_dir, container_runtime_installed):
+    container_kwargs = {
+        'process_isolation_executable': container_runtime_installed,
+        'process_isolation': True,
+        'container_image': 'quay.io/ansible/ansible-runner:devel'
+    }
+    private_data_dir = os.path.join(test_data_dir, 'debug')
+    inventory1 = os.path.join(private_data_dir, 'inventory', 'inv_1')
+    inventory2 = os.path.join(private_data_dir, 'inventory', 'inv_2')
+
+    out, _ = get_inventory(
+        action='list',
+        inventories=[inventory1, inventory2],
+        response_format='json',
+        quiet=True,
+        **container_kwargs
+    )
+    assert 'host_1' in out['ungrouped']['hosts']
+    assert 'host_2' in out['ungrouped']['hosts']
